[Fix: #108 ] : created separate file for ZOD schema's
