+++ src/app/(authentication)/auth/login/page.tsx
@@ -12,8 +12,9 @@ import { toast } from 'sonner';
 import { withAuthRedirect } from '@/HOCs/withAuthRedirect';
 import { useUsersApiAuthLoginUser } from '@/api/users-auth/users-auth';
 import Button from '@/components/common/Button';
-import FormInput, { emailOrUsernameSchema } from '@/components/common/FormInput';
+import FormInput from '@/components/common/FormInput';
 import { ArrowNarrowLeft } from '@/components/ui/Icons/common';
+import { emailOrUsernameSchema } from '@/constants/zod-schema';
 import { usePathTracker } from '@/hooks/usePathTracker';
 import { showErrorToast } from '@/lib/toastHelpers';
 import { useAuthStore } from '@/stores/authStore';
